+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelogs
 
+## 1.8.0 - July 28, 2018
+
+- Added: Exclude VSCode extensions from being synchronized.
+
+- Added: Automatically update your extensions during the synchronization.
+
+- Changed: Changed some settings of `Syncing` to improve the scalability.
+
+
 ## 1.7.3 - July 15, 2018
 
 - Changed: README.